requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.1.1=pypi_0
addict=2.4.0=pypi_0
blas=1.0=mkl
blinker=1.6.2=py310h06a4308_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.9.24=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2024.8.30=pyhd8ed1ab_0
cffi=1.17.1=py310h1fdaa30_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310h06a4308_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.3.1=pypi_0
cryptography=43.0.3=py310hdda0065_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
diffusers=0.30.3=pyhd8ed1ab_0
eigen=3.4.0=hdb19cb5_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py310h06a4308_0
flask=3.0.3=py310h06a4308_0
flask-login=0.6.3=py310h06a4308_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.10.0=pyhff2d567_0
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
harfbuzz=4.3.0=hf52aaf7_2
hdf5=1.12.1=h2b7332f_3
huggingface_hub=0.26.2=pyh0610db2_0
icu=73.1=h6a678d5_0
idna=3.7=py310h06a4308_0
importlib-metadata=8.5.0=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
itsdangerous=2.2.0=py310h06a4308_0
jinja2=3.1.4=py310h06a4308_1
jpeg=9e=h5eee18b_3
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.77=0
libcurl=8.9.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h77fa898_1
libgcc-ng=14.2.0=h69a702a_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=14.2.0=h77fa898_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hecde1de_4
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libopus=1.3.1=h5eee18b_1
libpng=1.6.39=h5eee18b_0
libpq=17.0=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libssh2=1.11.0=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.1=hfdd30dd_2
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py310h5eee18b_0
matplotlib=3.9.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mmcls=0.25.0=pypi_0
mmcv-full=1.7.2=pypi_0
mmdet=2.28.2=pypi_0
mmsegmentation=0.30.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multiscaledeformableattention=1.0=pypi_0
mysql=8.4.0=h29a9f33_1
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py310h06a4308_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
opencv=4.10.0=py310h0a8ef67_0
opencv-python=4.10.0.84=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openssl=3.4.0=hb9d3cd8_0
packaging=24.2=pyhff2d567_1
pcre2=10.42=hebb0a14_1
pillow=11.0.0=py310hfdbf927_0
pip=24.2=py310h06a4308_0
pixman=0.40.0=h7f8727e_1
platformdirs=4.3.6=pypi_0
prettytable=3.12.0=pypi_0
psutil=6.1.0=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pymysql=1.0.2=py310h06a4308_1
pyparsing=3.2.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.15=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pytorch=2.5.1=py3.10_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py310h5eee18b_0
qt-main=5.15.2=hb6262e9_11
readline=8.2=h5eee18b_0
regex=2024.11.6=py310ha75aee5_0
requests=2.32.3=py310h06a4308_1
safetensors=0.4.5=py310h505e2c1_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py310h06a4308_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
terminaltables=3.1.10=pypi_0
timm=1.0.11=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.20.3=pypi_0
tomli=2.1.0=pypi_0
torchaudio=2.5.1=py310_cu121
torchtriton=3.1.0=py310
torchvision=0.20.1=py310_cu121
tqdm=4.67.0=pyhd8ed1ab_0
transformers=4.46.3=pypi_0
typing-extensions=4.11.0=py310h06a4308_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py310h06a4308_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.6=py310h06a4308_0
wheel=0.44.0=py310h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yapf=0.43.0=pypi_0
zipp=3.21.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
